+++ turbo_transformers/python/turbo_transformers/layers/utils.py
@@ -1,0 +1,58 @@
+# Copyright (C) 2020 THL A29 Limited, a Tencent company.
+# All rights reserved.
+# Licensed under the BSD 3-Clause License (the "License"); you may
+# not use this file except in compliance with the License. You may
+# obtain a copy of the License at
+# https://opensource.org/licenses/BSD-3-Clause
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" basis,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+# implied. See the License for the specific language governing
+# permissions and limitations under the License.
+# See the AUTHORS file for names of contributors.
+
+import torch
+import torch.utils.dlpack as dlpack
+from typing import Union
+try:
+    # `turbo_transformers_cxxd` is the name on debug mode
+    import turbo_transformers.turbo_transformers_cxxd as cxx
+except ImportError:
+    import turbo_transformers.turbo_transformers_cxx as cxx
+from .return_type import convert_returns_as_type, ReturnType
+
+__all__ = [
+    'try_convert', 'convert2tt_tensor', 'to_param_dict_convert_tt',
+    'to_param_dict', 'create_empty_if_none', 'AnyTensor'
+]
+
+
+def convert2tt_tensor(t):
+    return cxx.Tensor.from_dlpack(dlpack.to_dlpack(t))
+
+
+def try_convert(t):
+    if isinstance(t, torch.Tensor):
+        return convert2tt_tensor(t)
+    elif isinstance(t, np.ndarray):
+        return convert2tt_tensor(torch.from_numpy(t))
+    else:
+        return t
+
+
+def to_param_dict_convert_tt(torch_module: torch.nn.Module):
+    return {
+        k: convert2tt_tensor(v)
+        for k, v in torch_module.named_parameters()
+    }
+
+
+def to_param_dict(torch_module: torch.nn.Module):
+    return {k: v for k, v in torch_module.named_parameters()}
+
+
+def create_empty_if_none(output):
+    return output if output is not None else cxx.Tensor.create_empty()
+
+
+AnyTensor = Union[cxx.Tensor, torch.Tensor]
